Moved chore pin board related files into dedicated folder.
