+++ client/src/parent/ParentScreen.tsx
@@ -2,7 +2,7 @@ import { DashboardLayout } from "../shell/DashboardLayout";
 import { Switch, Route, useRouteMatch } from "react-router-dom";
 import { ProfileDetails } from "./profile/ProfileDetails";
 import { RewardsList } from "./rewards/RewardsList";
-import { ChoreList } from "./ChoreList";
+import { ChoreList } from "./chores/ChoreList";
 
 export function ParentScreen() {
   const { path } = useRouteMatch();
