+++ client/src/parent/chores/ChoreCard.tsx
@@ -1,15 +1,15 @@
-import { ChoreWithAssignments } from "./models/ChoreWithAssignments";
+import { ChoreWithAssignments } from "../models/ChoreWithAssignments";
 import { Card, Typography } from "@material-ui/core";
 import { makeStyles, Theme, createStyles } from "@material-ui/core/styles";
 import { useState } from "react";
-import { AssignmentState } from "./models/AssignmentState";
+import { AssignmentState } from "../models/AssignmentState";
 import { AssignmentList } from "./AssignmentList";
-import { useInfoNotification } from "../Snackbar";
-import { MoreOptionsMenu } from "./MoreOptionsMenu";
-import { AddOptionsExpandCardActions } from "./AddOptionsExpandCardActions";
-import { BiAvatarCardHeader } from "./BiAvatarCardHeader";
-import { CollapsibleCardContent } from "./CollapsibleCardContent";
-import { AddButtonWithLabel } from "./AddButtonWithLabel";
+import { useInfoNotification } from "../../Snackbar";
+import { MoreOptionsMenu } from "../MoreOptionsMenu";
+import { AddOptionsExpandCardActions } from "../AddOptionsExpandCardActions";
+import { BiAvatarCardHeader } from "../BiAvatarCardHeader";
+import { CollapsibleCardContent } from "../CollapsibleCardContent";
+import { AddButtonWithLabel } from "../AddButtonWithLabel";
 
 type Prop = {
   chore: ChoreWithAssignments;
